requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
python-dotenv
numpy
pydantic
typing-extensions
faiss-cpu
rank-bm25
tiktoken
pypdf
rank_bm25